progress: 8.37% of code matched (739/8216 functions, 797/1245 objects)

+++ src/game/game_fn_8000E84C.c
@@ -22,8 +22,8 @@ extern void fn_801A7518(void*, s32);
 extern void fn_801A7538(void*, u16);
 extern void fn_801A7588(void*, s32);
 extern void fn_801A764C(void*, Vec3*);
-extern void fn_8020104C(int, void*, void*, int, float);
-#define fn_8020104C(a, b, c, d, e) fn_8020104C((a), (void*)(b), (void*)(c), (d), (e))
+extern void fn_8020104C(int, void *, void *, int, float);
+#define fn_8020104C(a, b, c, d, e) fn_8020104C((a), (void *)(b), (void *)(c), (int)(d), (e))
 
 void fn_8000E84C(s32 object_id, s32 value)
 {
