+++ src/game/game_fn_80025F48.c
@@ -24,8 +24,8 @@ extern void* fn_80204844(void*, int);
 extern int fn_80201B54();
 extern void *fn_80201B8C();
 extern unsigned long long fn_8020123C();
-extern void fn_8020104C(int, void*, void*, int, float);
-#define fn_8020104C(a, b, c, d, e) fn_8020104C((a), (void*)(b), (void*)(c), (d), (e))
+extern void fn_8020104C(int, void *, void *, int, float);
+#define fn_8020104C(a, b, c, d, e) fn_8020104C((a), (void *)(b), (void *)(c), (int)(d), (e))
 
 void fn_80025F48(void)
 {
