+++ src/game/game_fn_8002BC7C.c
@@ -22,14 +22,14 @@ extern void *fn_80201BC8();
 extern void fn_8011FA8C(void*, int, int);
 extern void fn_8011F788(void*, float);
 extern void fn_8011F798(void*, float);
-extern void fn_8020104C(int, void*, void*, int, float);
-#define fn_8020104C(a, b, c, d, e) fn_8020104C((a), (void*)(b), (void*)(c), (d), (e))
+extern void fn_8020104C(int, void *, void *, int, float);
+#define fn_8020104C(a, b, c, d, e) fn_8020104C((a), (void *)(b), (void *)(c), (int)(d), (e))
 extern unsigned long long fn_8020123C();
 extern void fn_80201D2C(void *, s32);
 extern void fn_80201D14(void *, s32);
 extern int fn_801E8328();
-extern void fn_80201D34(void*, s32);
-extern void fn_80201D1C(void*, s32);
+extern void fn_80201D34(void *, s32);
+extern void fn_80201D1C(void *, s32);
 extern void fn_8015424C(s32, s32, s32, s32, s32, s32, s32, s32);
 extern void fn_8002BFE0(void*);
 extern s32 fn_8015E4E8(void);
