+++ src/game/game_fn_8003232C.c
@@ -55,16 +55,16 @@ extern void *fn_80201814();
 extern int fn_80201B4C(void*);
 extern void fn_80201D2C(void *, s32);
 extern void fn_80201D14(void *, s32);
-extern void fn_80201D34(void*, s32);
-extern void fn_80201D1C(void*, s32);
+extern void fn_80201D34(void *, s32);
+extern void fn_80201D1C(void *, s32);
 extern int fn_80035628(void*);
 extern int fn_801D3974(s32);
 extern s32 fn_801D3A24(s32, s32);
 extern void fn_8014CFF4(void*, void*, s32, s32, void*, void*, s32);
 extern void fn_8014CE98(void*, void*, s32, s32, void*, s32, s32, s32, s32,
                        void*, void*, s32, s32, s32);
-extern void fn_8020104C(int, void*, void*, int, float);
-#define fn_8020104C(a, b, c, d, e) fn_8020104C((a), (void*)(b), (void*)(c), (d), (e))
+extern void fn_8020104C(int, void *, void *, int, float);
+#define fn_8020104C(a, b, c, d, e) fn_8020104C((a), (void *)(b), (void *)(c), (int)(d), (e))
 extern void fn_80191204(void*, s32);
 extern void fn_801911FC(void*, s32);
 extern u8 fn_80180130(void*);
