+++ src/game/game_fn_80060840.c
@@ -9,7 +9,7 @@ typedef enum Result {
 } Result;
 extern int fn_80200C38();
 extern u16 fn_801A7434(void *);
-extern u32 fn_800FBFB0(void);
+extern int fn_800FBFB0(void);
 
 Result fn_80060840(void *owner, s32 *special)
 {
@@ -30,8 +30,8 @@ Result fn_80060840(void *owner, s32 *special)
         *special = 1;
         break;
     default:
-        result = fn_800FBFB0() % 2 ? ResultFour :
-                 (fn_800FBFB0() % 2 ? ResultFive : ResultSix);
+        result = (u32)fn_800FBFB0() % 2 ? ResultFour :
+                 ((u32)fn_800FBFB0() % 2 ? ResultFive : ResultSix);
         break;
     }
